renamed retry graph to a more appropriate class name
